FIX: ReductionOptions issues when reducing
Needed to mess around with ReductionOptions to tell refnx when to reduce m_spec and when to reduce m_spec_polcorr. Still not 100% sure whether this is foolproof, but I am able to plot both raw and polarisation-corrected data using this. Need to make more robust for less vanilla test cases.

Also included new functions in __init__ file