Added warning and method to mask zeros in Data1D
When creating objects that inherit from Data1D, a warning is raised if the y or y_err data contain zeros. This is due to the NaNs that can result in Objective.logl if the data is left as is. A method to mask these data points has also been added.