ADD: start of PolarisedReduce class and SpinSet class
PolarisedReduce encompasses all data for a typical reduction of polarised neutron reflectometry data. It needs direct beams for initialisation, that is made into a SpinSet which creates a PlatypusReduce object for each spin channel measured. TODO: add reflected beams into object for reduction

SpinSet is an class to keep a spin set for a given angle together. it creates a PLatypusReduce object for each spin channel. TODO: incorporate polarisation correction in the initialisation routine

Added both classes to __init__.py