[BUG] Force event-mode logs to be read as LE int32
Fix event data processing on big-endian architechtures:
The event-mode files in the data bundle for the tests are always big-endian
as they are generated on the insturment machines. Opening the files needs
to be specific about the byte-order and not open with an implicitly
native byte-order as np.frombuffer normally does.